add requirement for 'shift' before triggering a click for click delta (#172)
* also plot with line_width for raw plots

* Allow highlighting of plots from 2 logs by having the settings/metadata from one log open and hovering on another ones name

+++ crates/plot_util/src/plots/plot_data.rs
@@ -43,7 +43,7 @@ impl PlotData {
     }
 
     /// Adds a plot to the [`PlotData`] collection if another plot with the same label doesn't already exist
-    pub fn add_plot_if_not_exists(&mut self, raw_plot: &RawPlot, log_id: usize) {
+    pub fn add_plot_if_not_exists(&mut self, raw_plot: &RawPlot, log_id: u16) {
         let mut plot_label = String::with_capacity(30); // Approx. enough to not reallocate
         plot_label.push('#');
         plot_label.push_str(&log_id.to_string());
@@ -77,7 +77,7 @@ pub struct PlotValues {
     mipmap_max: MipMap2D<f64>,
     mipmap_min: MipMap2D<f64>,
     name: String,
-    log_id: usize,
+    log_id: u16,
     // Label = "<name> #<log_id>"
     label: String,
     color: Color32,
@@ -90,7 +90,7 @@ impl PlotValues {
     // Don't mipmap/downsample to more than this amount of elements
     const MIPMAP_MIN_ELEMENTS: usize = 512;
 
-    pub fn new(raw_plot: Vec<[f64; 2]>, name: String, log_id: usize) -> Self {
+    pub fn new(raw_plot: Vec<[f64; 2]>, name: String, log_id: u16) -> Self {
         let label = format!("{name} #{log_id}");
         Self {
             mipmap_max: MipMap2D::without_base(
@@ -192,7 +192,7 @@ impl PlotValues {
     }
 
     /// The ID of the log that the plot belongs to
-    pub fn log_id(&self) -> usize {
+    pub fn log_id(&self) -> u16 {
         self.log_id
     }
 
@@ -215,13 +215,13 @@ impl PlotValues {
 /// Represents all the plotlabels from a given log
 #[derive(Debug, PartialEq, Deserialize, Serialize)]
 pub struct StoredPlotLabels {
-    pub log_id: usize,
+    pub log_id: u16,
     pub label_points: Vec<PlotLabel>,
     pub highlight: bool,
 }
 
 impl StoredPlotLabels {
-    pub fn new(label_points: Vec<([f64; 2], String)>, log_id: usize) -> Self {
+    pub fn new(label_points: Vec<([f64; 2], String)>, log_id: u16) -> Self {
         Self {
             label_points: label_points.into_iter().map(PlotLabel::from).collect(),
             log_id,
@@ -243,7 +243,7 @@ impl StoredPlotLabels {
         self.label_points.iter_mut().map(|label| &mut label.point)
     }
 
-    pub fn log_id(&self) -> usize {
+    pub fn log_id(&self) -> u16 {
         self.log_id
     }
